+++ apps/desktop/electron.vite.config.1780887814998.mjs
@@ -1,0 +1,42 @@
+// electron.vite.config.ts
+import { resolve } from "path";
+import { defineConfig } from "electron-vite";
+import react from "@vitejs/plugin-react";
+var __electron_vite_injected_dirname = "/sessions/charming-intelligent-ritchie/mnt/outline-desktop-client/apps/desktop";
+var electron_vite_config_default = defineConfig({
+  main: {
+    build: {
+      outDir: "out/main",
+      rollupOptions: {
+        input: {
+          index: resolve(__electron_vite_injected_dirname, "src/main/index.ts")
+        }
+      }
+    }
+  },
+  preload: {
+    build: {
+      outDir: "out/preload",
+      rollupOptions: {
+        input: {
+          index: resolve(__electron_vite_injected_dirname, "src/preload/index.ts")
+        }
+      }
+    }
+  },
+  renderer: {
+    root: resolve(__electron_vite_injected_dirname, "src/renderer"),
+    build: {
+      outDir: resolve(__electron_vite_injected_dirname, "out/renderer"),
+      rollupOptions: {
+        input: {
+          index: resolve(__electron_vite_injected_dirname, "src/renderer/index.html")
+        }
+      }
+    },
+    plugins: [react()]
+  }
+});
+export {
+  electron_vite_config_default as default
+};